+++ scripts/mvn-site.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+BASEDIR="$(cd "$(dirname "$0")" && pwd -P)"/..
+# Provides the logi function
+source "$BASEDIR"/scripts/logger.sh
+MVN="$BASEDIR"/mvnw
+
+logi "Build Maven site"
+"$MVN" site
+"$MVN" site:stage